src/ska-tmc-dishleafnode-mid/requirements.txt
numpy == 1.17.2
pytango >= 9.3.3
jsonschema >= 3.2.0
ska-ser-logging
marshmallow
ska-tmc-cdm == 6.0.0
fire
ska-tmc-common==0.1.7+d39e6423
ska-ser-log-transactions
astropy>=4.1
katpoint==1.0a1
datetime
ska-tango-base == 0.7.2